Improve copying text selection to clipboard when page text layer not loaded and add documentation
